FieldToEntityColumnConverter
